fix(funnel): adjust import paths post src-layer consolidation rebase
After rebasing onto develop (post #163 src/ → src/{studio,modeling,…}
consolidation), funnel route files and FunnelViewer reference paths
that no longer exist:

- src/studio/routes/*.tsx imports from '../funnel/...' need '../../funnel/...'
  (routes moved from src/routes to src/studio/routes)
- src/funnel/components/FunnelViewer imports '../../components/Viewer'
  and '../../context/WorkbenchContext'; both moved to src/studio/

Pure path-rewrite, no behavior change. Typecheck PASS.
